Move extract method to utils ++
